--- a/src/simulat/core/surfaces/scenes/main_menu_scene/main_menu_scene.py
+++ b/src/simulat/core/surfaces/scenes/main_menu_scene/main_menu_scene.py
@@ -53,6 +53,7 @@ class MainMenuScene(Scene):
         # add logo
         logo_path = "assets/logo/ingame/logo_menu.png"
         self.logo = pg.image.load(logo_path).convert_alpha()
+        self.logo = pg.transform.scale(self.logo, (412, 82))
 
     def _load_game(self):
         from src.simulat.core.surfaces.scenes.game_scene.game_scene import \
@@ -99,7 +100,7 @@ class MainMenuScene(Scene):
         simulat.topbar.update_title("main menu")
 
         # draw logo
-        self.surface.surface.blit(self.logo, (240, 200))
+        self.surface.surface.blit(self.logo, (120, 100))
 
         # draw buttons
         self.button_container.render()