… assets/fonts/simulat.ttf
… assets/logo/banner/banner.png
… assets/logo/banner/banner_big.png
… assets/logo/ingame/logo_menu.png
… assets/logo/logo.png
… assets/logo/logo.svg