## PR type (check all applicable)

- [ ] `   feat   ` :sparkles: features
- [ ] `   fix    ` :bug: bugfixes
- [ ] `  chore   ` :ticket: chores
- [ ] `refactor` :package: code refactoring
- [ ] `  style   ` :gem: style
- [ ] `   docs   ` :book: documentation changes
- [ ] `   perf   ` :rocket: performance improvements
- [ ] `   test   ` :rotating_light: tests
- [ ] `  build   ` :construction_worker: build
- [ ] `    ci    ` :robot: continuous integration
- [ ] `  revert  ` :back: reverts

## PR description

This PR:
- Moves `src/simulat/assets` to `assets/` (project root).
- #55 
  - Reduces internal rendering resolution to 640x360 and scales to display resolution (default: 1280x720)
- #56 
  - Adds the simulat font.

## Related Tickets

- related issue #55, #56
- closes #

## Instructions, Screenshots

### #55 (scaling)

Peformance improvements:

| Revision                            | Main Menu  | Game Map   |
|-------------------------------------|------------|------------|
| Before (1280x720 native)            | 498.26 FPS | 296.80 FPS |
| After (640x360, scaled to 1280x720) | 572.10 FPS | 430.14 FPS |
| Performance improvement             | +14.92%    | +44.92%    |

Graphical comparison:

Scaled:
![scaling_640x360_new](https://github.com/pufereq/simulat/assets/94570596/13aeba89-0384-4fc2-955d-d3dcc0243257)

Native:
![scaling_old_0.13.0](https://github.com/pufereq/simulat/assets/94570596/3db21429-dbf0-480f-b544-fb81c06cdf7f)

### #56 (font)

Main menu:
![a screenshot of main menu with new font](https://github.com/pufereq/simulat/assets/94570596/2ae9f4e0-aeec-4633-83e2-7504b78e0b3a)

Loading text:
![a screenshot of game map loading text](https://github.com/pufereq/simulat/assets/94570596/511dbfdd-89fd-424e-9848-e4fc3d210ab0)

Part of the game scene's UI:
![a screenshot of a part of game scene's UI](https://github.com/pufereq/simulat/assets/94570596/21821cee-76db-488b-9f8e-59401a99d531)

simulat font specimen:
![simulat font specimen image](https://github.com/pufereq/simulat/assets/94570596/8eb04b03-1fc6-45dd-90c9-d851e858814f)
_Generated with: http://shell.aiei.ch/typography/_